pyproject.toml
[project]
name = "db-mvp"
version = "0.1.0"
description = "Brand Intelligence API with Streamlit Frontend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic",
    "supabase",
    "python-dotenv",
    "streamlit",
    "requests",
    "pandas",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["apis.py", "main.py"]